# client/openapi/rudoger-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Rudoger.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
...
    "/api/v1/inventory/stock-items": {
      "post": {
        "tags": [
          "StockItems"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "StockItems"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfStockItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfStockItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfStockItemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/inventory/stock-items/{stockItemId}": {
      "get": {
        "tags": [
          "StockItems"
        ],
        "parameters": [
          {
            "name": "stockItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/inventory/stock-items/{stockItemId}/movements": {
      "post": {
        "tags": [
          "StockItems"
        ],
        "parameters": [
          {
            "name": "stockItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockMovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StockMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockMovementResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "StockItems"
        ],
        "parameters": [
          {
            "name": "stockItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfStockMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfStockMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfStockMovementResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/order/orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlacementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlacementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlacementResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfOrderResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfOrderResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseOfOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/order/orders/{orderId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/order/order-placements/{placementId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "placementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlacementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlacementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlacementResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/order/orders/{orderId}/transitions": {
      "post": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderTransitionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderTransitionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderTransitionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderTransitionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderTransitionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/order/orders/{orderId}/transitions/{transitionId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "transitionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderTransitionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderTransitionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderTransitionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateOrderRequest": {
        "examples": [
          {
            "lines": [
              {
                "productId": "00000000-0000-0000-0000-000000000000",
                "uomCode": "CASE",
                "quantity": 2
              }
            ]
          }
        ],
        "required": [
          "lines"
        ],
        "type": "object",
        "properties": {
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderPlacementLineRequest"
            }
          }
        }
      },
      "CreateOrderTransitionRequest": {
        "examples": [
          {
            "target": "Shipped"
          }
        ],
        "required": [
          "target"
        ],
        "type": "object",
        "properties": {
          "target": {
            "$ref": "#/components/schemas/OrderTransitionTargetContract"
          }
        }
      },
...
      "CreateStockItemRequest": {
        "examples": [
          {
            "productId": "00000000-0000-0000-0000-000000000000",
            "uomCode": "CASE",
            "openingQuantity": 10
          }
        ],
        "required": [
          "productId",
          "uomCode",
          "openingQuantity"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "uomCode": {
            "type": "string"
          },
          "openingQuantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "CreateStockMovementRequest": {
        "examples": [
          {
            "type": "Receipt",
            "uomCode": "CASE",
            "quantity": 2
          }
        ],
        "required": [
          "type",
          "uomCode",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/StockMovementTypeContract"
          },
          "uomCode": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "OrderPlacementLineRequest": {
        "required": [
          "productId",
          "uomCode",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "uomCode": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "OrderPlacementLineResponse": {
        "required": [
          "num",
          "productId",
          "uomCode",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "num": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "uomCode": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "OrderPlacementResponse": {
        "required": [
          "id",
          "orderId",
          "userId",
          "status",
          "lines",
          "failureCode",
          "failureDetail"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/OrderPlacementStatusContract"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderPlacementLineResponse"
            }
          },
          "failureCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "failureDetail": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "OrderPlacementStatusContract": {
        "type": "integer"
      },
      "OrderResponse": {
        "required": [
          "id",
          "orderNumber",
          "status",
          "userId",
          "lines",
          "pendingTransitionId",
          "pendingTransitionTarget"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatusContract"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineResponse"
            }
          },
          "pendingTransitionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "pendingTransitionTarget": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OrderTransitionTargetContract"
              }
            ]
          }
        }
      },
      "OrderStatusContract": {
        "type": "integer"
      },
      "OrderTransitionResponse": {
        "required": [
          "id",
          "orderId",
          "target",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "target": {
            "$ref": "#/components/schemas/OrderTransitionTargetContract"
          },
          "status": {
            "$ref": "#/components/schemas/OrderTransitionStatusContract"
          }
        }
      },
...
      "PageResponseOfOrderResponse": {
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PageResponseOfStockItemResponse": {
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockItemResponse"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "PageResponseOfStockMovementResponse": {
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockMovementResponse"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "StockItemResponse": {
        "required": [
          "id",
          "productId",
          "baseUomCode",
          "onHandQuantity",
          "reservedQuantity",
          "availableQuantity"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "baseUomCode": {
            "type": "string"
          },
          "onHandQuantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "reservedQuantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "availableQuantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "StockMovementResponse": {
        "required": [
          "id",
          "stockItemId",
          "type",
          "uomCode",
          "quantity",
          "onHandQuantityDelta",
          "reservedQuantityDelta",
          "referenceType",
          "referenceId",
          "idempotencyKey",
          "correlationId",
          "occurredAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "stockItemId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/StockMovementTypeContract"
          },
          "uomCode": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "onHandQuantityDelta": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "reservedQuantityDelta": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "referenceType": {
            "type": "string"
          },
          "referenceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "occurredAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StockMovementTypeContract": {
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token issued by POST /api/v1/authn/tokens.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}